refactor(checker): split type environment location helpers
